+++ src/namdesktop/ui/MainFrame.java
@@ -1,15 +1,17 @@
 package namdesktop.ui;
 
+import namdesktop.model.NamWorkspace;
+
 import javax.swing.*;
 import java.awt.*;
 
 public final class MainFrame extends JFrame {
 
-    public MainFrame() {
-        var leftPanel = new JPanel();
+    public MainFrame(NamWorkspace workspace) {
+        var treePanel = new TreePanel(workspace);
         var centrePanel = new JPanel();
 
-        var splitPane = new JSplitPane(JSplitPane.HORIZONTAL_SPLIT, leftPanel, centrePanel);
+        var splitPane = new JSplitPane(JSplitPane.HORIZONTAL_SPLIT, treePanel, centrePanel);
         splitPane.setDividerLocation(220);
         splitPane.setResizeWeight(0.0);